Composite property tree nodes for will-change only when relevant property has will-change.
This causes the test
external/wpt/css/filter-effects/effect-reference-feimage-002.html to
start failing (in the same way as -001 and -003) because the filter
effect is no longer composited.  It creates a copy of the test as -004
with will-change: filter rather than will-change: transform that
continues to have a composited filter effect node and thus continues
passing.

Bug: 900241
Change-Id: I18efc630668ea7eac2a8c0a0417b766762e71d9f