Ensure backdrop filter cc::Layers are marked as drawing content
ContentLayerClientImpl::UpdateCcPictureLayer needs to set
cc::Layer::DrawsContent to true for content under a backdrop filter.
The previous code would do this based on the PendingLayer's effect node,
but this is incorrect if a single object has both a backdrop filter
effect node and a filter effect node. This approach failed to re-run
UpdateCcPictureLayer when HasBackdropEffect changed (see new WPT test
repaint-added-backdrop-filter.html).

This patch moves the source-of-truth for draws-content into PaintChunk,
and sets it based on whether any effect node could draw content.

Bug: 1273849
Change-Id: Ibb7c15361e83310881311122bf2cbec479a43d3a