Add options parameter to createSyncAccessHandle()
For now the only accepted (and required) option is "mode: 'in-place'".
Having this parameter allows us to later define a more appropriate
default behavior for access handles than the one we currently implement.
This is specially useful in the context of expanding access handle usage
outside of OPFS.

For more details check:
https://github.com/WICG/file-system-access/blob/main/AccessHandle.md#exposing-accesshandles-on-all-filesystems

Bug: 1218431
Change-Id: I50d0a4acfd81874e17b6b28d1c2bf446b0397a4d