CSSMathInvert should serialize 1/0 expression
According to the spec [1], the expression of 1/0 should be serialized.

However, In DetermineCategory() in css_math_expression_node.cc
evaluates kCalcOther.

Therefore, Evaluate conditionally the division by zero with
a feature flag 'Allow Infinity and NaN in CSS calc' to prevent
return nullptr in the creation of the expression tree.

[1] https://www.w3.org/TR/css-typed-om-1/#calc-serialization

Bug: 1240306
Change-Id: I3c7705665af71d47ef3140bdd5671569dc394da4