Refactor srcdoc to use navigation params, fix reset.
This CL removes the assumption that the owner of a srcdoc iframe lives
in the same process as the iframe's comment. It does this by sending
the srcdoc content to the browser in a mojo IPC, and sends it back
to the renderer via CommitNavigationParams. The content value is also
stored on the FrameTreeNode, and as such will be available in the
history.

As a byproduct of this refactor, the test
http/tests/navigation/srcdoc-attribute-reset.html appears to fail, but
it was actually flawed in the first place, since it never verified that
the srcdoc content was rendered before the attribute was reset. In fact
the content was being cleared before it ever rendered, allowing the
test to pass, even though manually resetting srcdoc shows that the
content doesn't get cleared.

This CL fixes the test (and converts it to a WPT test), and makes sure
the test passes by detecting the srcdoc's value being cleared and
navigating the iframe to the blank url if no src attribute is available
to use instead.

Bug: 1233143
Change-Id: Ie511980dc6dfa4ec4cea18e4b1baf78c66a2a420